Rename alma namespace into alma_client.
