+++ alma_client/entities/base.py
