+++ alma_client/entities/eligibility.py
