+++ alma_client/entities/merchant.py
