+++ tox.ini
@@ -1,10 +1,10 @@
 [tox]
-envlist = py36,py37,py38,py39,flake8
+envlist = py36,py37,py38,py39,py310,flake8
 
 [testenv]
 commands =
-    mypy --ignore-missing-imports --scripts-are-modules alma
-    pytest tests --cov-report term-missing --cov alma {posargs} # --cov-fail-under 100
+    mypy --ignore-missing-imports --scripts-are-modules alma_client
+    pytest tests --cov-report term-missing --cov alma_client {posargs} # --cov-fail-under 100
 deps =
     -rtest-requirements.txt
 	-rdev-requirements.txt
@@ -12,11 +12,11 @@ install_command = pip install {opts} {packages}
 
 [testenv:py35]
 commands =
-    pytest tests --cov-report term-missing --cov alma {posargs} # --cov-fail-under 100
+    pytest tests --cov-report term-missing --cov alma_client {posargs} # --cov-fail-under 100
 deps = -rtest-requirements.txt
 
 [testenv:flake8]
-commands = flake8 alma
+commands = flake8 alma_client
 deps =
     flake8
